split model into independent files
