+++ swan/models/__init__.py
@@ -1,1 +1,6 @@
 """Models API."""
+
+from .fingerprint_models import FingerprintFullyConnected
+from .graph_models import ChemiNet
+
+__all__ = ["ChemiNet", "FingerprintFullyConnected"]
