+++ swan/models/fingerprint_models.py
@@ -1,0 +1,24 @@
+"""Statistical models."""
+
+from torch import Tensor, nn
+
+__all__ = ["FingerprintFullyConnected"]
+
+
+class FingerprintFullyConnected(nn.Module):
+    """Fully connected network for non-linear regression."""
+
+    def __init__(self, n_feature: int, n_hidden: int):
+        """Create a deep feed foward network."""
+        super().__init__()
+        self.seq = nn.Sequential(
+            nn.Linear(n_feature, n_hidden),
+            nn.ReLU(),
+            nn.Linear(n_hidden, n_hidden),
+            nn.ReLU(),
+            nn.Linear(n_hidden, 1),
+        )
+
+    def forward(self, tensor: Tensor) -> Tensor:
+        """Run the model."""
+        return self.seq(tensor)
